crossbar: config: set user_timeout in default config
In crossbar v22.1.1 the "user_timeout" option was added [1], describing
it as follows:

  "Optional TCP user timeout in ms transmitted data may remain unacked
  before connection close, see TCP_USER_TIMEOUT in tcp(7), 0 behaves
  like not configured (default: not configured)" [2]

If a machine running labgrid-exporter is shut down forcibly, the
connection hangs and subsequent procedure registrations fail until the
connection is closed. This can take up to 20 minutes in the worst case.

This option limits the timespan the connection can hang.

[1] https://github.com/crossbario/crossbar/pull/1930
[2] https://crossbar.io/docs/Transport-Endpoints/#tcp-listening-endpoints

Signed-off-by: Bastian Krause <bst@pengutronix.de>