pyproject.toml/.pylintrc: add pylint config to pyproject.toml in favor of .pylintrc
The pylint results have never been considered because the load of
messages is pretty high. This is why the CI jobs never ran pylint.

On the other hand, pylint can reveal important issues. These issues
should be addressed before a PR is merged. In the past, this was done on
some PRs manually.

In order to improve the situation, stick to a basic set of relevant
messages. No styling messages are included, since we want to integrate
black [1] in a future PR. Since we want to head into the pyproject.toml
direction in the future anyway, add the relevant pylint configuration
there.

A later commit will let pylint run in our CI workflow.

[1] https://black.readthedocs.io/en/stable/

Signed-off-by: Bastian Krause <bst@pengutronix.de>