treewide: replace autobahn and crossbar with gRPC
Crossbar and autobahn are used to communicate between client and
coordinator as well as between exporter and coordinator. Both are
unfortunately not very well maintained anymore. The crossbar component
was moved to its own virtualenv for quite a while to cope with
conflicting dependencies. python3.13 support for crossbar is still not
available (at least not in a version releases on PyPi).

That's why labgrid will now move to gRPC. It's a well maintained RPC
framework. As a side effect, the message transfer is more performant and
the import times are shorter.

gRPC relies on the protocol buffers compiler (protoc) to generate code.
Use grpcio-tools 1.62.2 to do that. It's used in Yocto scarthgap and is
incompatible with generated code from newer grpcio-tools.

Co-developed-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>
Co-developed-by: Bastian Krause <bst@pengutronix.de>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>