remote/coordinator: refactor handling of orphaned resources
If an exporter disconnects, corresponding resources that are acquired
become orphaned. If the exporter reconnects, the resources are not
locked. Requiring the user to unlock and lock the corresponding place
again is inconvenient and would change the previous behavior.

Prior to this commit, reacquiring did not work due to a logic error:
ExporterCommand.complete() und ExporterCommand.wait() are both called in
ExporterStream.request_task(). The blocking wait() prevents further
processing of exporter messages. That also means responses for
ExporterSetAcquiredRequest are not handled anymore. This leads to a
state where resources cannot be acquired/released by their place anymore.
Ultimatively, this leads to an inconsistent state requiring a
coordinator restart.

Refactor handling of these orphaned resources to solve this.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>