driver/rawnetworkinterfacedriver: add interface up/down/wait_state
Add methods to the RawNetworkInterfaceDriver to set interfaces up and
down, as well as getting and waiting for the interface state.

This allows the driver to take more control over the interface,
preconfiguration is not needed anymore. Tests that expect the exporter
interface to be down (such as ethernet selftests, cable tests) are now
possible.

Note that the RawNetworkInterfaceDriver now brings the bound interface
up on activate and down on deactivate.

Signed-off-by: Bastian Krause <bst@pengutronix.de>