remote/client: provide default for ClientSession.env to make it optional
The Environment was always optional. Before users of ClientSession had
to pass an explicit None for this attribute. While we're changing the
ClientSession for gRPC anyway, let's make env really optional.

This again allows us to make start_session()'s extra argument optional,
too. It is used to pass extra arguments to the ClientSession, which
means it can be an empty dictionary now.

Signed-off-by: Bastian Krause <bst@pengutronix.de>