remote/exporter: make coordinator port default to 20408
The coordinator binds to port 20408 by default. The client tries to
connect to that port if no port is specified via --cordinator/-x,
LG_COORDINATOR or via coordinator_address option in the environment
config.

Only the exporter does not yet default to 20408. Change that by
appending the default port 20408 to the address given via
--coordinator/-c or LG_COORDINATOR if no port was specified.

Closes #1429

Signed-off-by: Bastian Krause <bst@pengutronix.de>