helpers/labgrid-raw-interface: make tcpdump count argument actually optional
The intention of labgrid-raw-interface was to have an optional "count"
argument, allowing for packet capture on an interface until a timeout is
reached, no matter how many packets are received. This is used when
calling:

  RawNetworkInterfaceDriver.start_record(filename, timeout=10)

However, the missing `nargs="?"` kwarg when adding the argument actually
required the argument to be given:

  usage: labgrid-raw-interface tcpdump [-h] [--timeout TIMEOUT] ifname count
  labgrid-raw-interface tcpdump: error: the following arguments are
  required: count

The original commit introducing the option actually included it [1]. Fix
that by re-adding `nargs="?"`.

[1] 90cf9413 ("helpers: add labgrid-raw-interface helper")

Fixes: c1aec37e ("helpers/labgrid-raw-interface: introduce subparsers per program, pass options")
Signed-off-by: Bastian Krause <bst@pengutronix.de>
(cherry picked from commit 1e207b22b4a58355960863d593c8d759c176c925)