remote/exporter: Export invert parameter on relays
The invert parameter is currently not being exported on Network HID and
Deditec relays. Since the client is the one responsible for inverting
the signal, and it does not recieve this information from the exporter
(which causes it to default to False), the invert is only working on
local setups and not on network exported devices.

Signed-off-by: Diogo Silva <diogo.silva.pt14@gmail.com>
(cherry picked from commit 053fbf01c6bb52e53fedea19fe2cf4c78caa5a21)